requirements.txt
streamlit
groq
langchain
langchain-community
langchain-groq
faiss-cpu
python-dotenv
tf-keras
sentence-transformers